Use yesterday's delta if today's delta is 0
It seems that sometimes John Hopkins has trouble
collecting the values so yesterday's values show up
in csvs causing a delta of 0 which gives all green
circles. To smooth this over I'm now reusing yesterday's
delta if today's is 0. This will smooth over only 1 day.
If this will not be enough I'll try and come up with
something more sophisticated.