Adapt to change in source data format
John Hopkins changed the names of files and also changed the
data in csvs a bit. Recovered numbers are no longer being collected
and it has data for some locations not in confirmed csv.
This commit adapts to new source data structure.

What's missing is estimating recovered cases because new data
for recovered will no longer be compiled by John Hopkins. This
will come in a subsequent commit.